requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
geographiclib==1.52
geopy==2.2.0
factory-boy==3.2.1
Faker==15.1.0
idna==3.4
numpy==1.23.4
pandas==1.5.1
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
